backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["gudezhi <gudezhi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
gevent = "^23.9.1"
drf-yasg = "^1.21.7"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
